+++ R/prepare_models.R
@@ -23,7 +23,7 @@
 read_model_controls <- function (main     = ".",
                                  settings = directory_settings()) {
 
-  read_yaml(file = file.path(main, settings$subs$models, settings$files$model_controls))
+  read_yaml(file = file.path(main, settings$subdirectories$models, settings$files$model_controls))
 
 }
 
@@ -81,7 +81,7 @@ write_model_controls <- function (main               = ".",
   }
 
   write_yaml(x    = model_controls,
-             file = file.path(main, settings$subs$models, settings$files$model_controls))
+             file = file.path(main, settings$subdirectories$models, settings$files$model_controls))
 
   invisible(model_controls)
 
@@ -143,7 +143,7 @@ write_model <- function (main     = ".",
   }
 
   model_file <- paste0(model, ".R")
-  mod_path   <- file.path(main, settings$subs$models, model_file)
+  mod_path   <- file.path(main, settings$subdirectories$models, model_file)
 
 
   mod_template <- model_template(main     = main,